Finished main routes for backend

+++ backend/routes/users.js
@@ -2,6 +2,8 @@ const express = require("express");
 const mongoose = require("mongoose");
 const saveImage = require("../utils/saveImage");
 const fs = require("fs");
+const createUser = require("../utils/createUser");
+const deleteUser = require("../utils/deleteUser");
 
 const { User, validateUser } = require("../models/user");
 const router = express.Router();
@@ -11,6 +13,7 @@ router.get("/", async (req, res) => {
   const users = await User.find({}).sort("name");
   res.send(users);
 });
+
 // get user by id
 router.get("/:id", async (req, res) => {
   const id = req.params.id;
@@ -20,6 +23,45 @@ router.get("/:id", async (req, res) => {
   if (user) return res.send(user);
   res.status(404).send("User not found");
 });
+
+// create user
+router.post("/", async (req, res) => {
+  const { userBody, otherBody } = req.body;
+
+  const errorUser = validateUser(userBody);
+  if (errorUser) res.status(400).send(errorUser);
+
+  const exist = await User.findOne({ email: userBody.email });
+  if (exist)
+    return res.status(400).send("User already exists with existing email");
+
+  const user = new User(userBody);
+  otherBody.user = user._id;
+
+  const other = createUser(user.role, otherBody);
+  if (other.errorBody) return res.status(400).send(other.errorBody);
+
+  // starting session to perform task set
+  // if anything goes wrong tasks will rollback
+  const session = await User.startSession();
+
+  try {
+    // add session parameter for each save({session})
+    // transactions only supported by replica set
+    await user.save();
+    await other.body.save();
+
+    res.send(other);
+
+    await session.commitTransaction();
+  } catch (err) {
+    res.status(400).send(err.message);
+    await session.abortTransaction();
+  } finally {
+    session.endSession();
+  }
+});
+
 // update user by id
 router.put("/:id", async (req, res) => {
   const id = req.params.id;
@@ -34,8 +76,9 @@ router.put("/:id", async (req, res) => {
 
   res.status(404).send("User not found");
 });
+
 // update user image
-router.patch("/:id", async (req, res) => {
+router.patch("/picture/:id", async (req, res) => {
   const picture = req.files ? req.files.picture || undefined : undefined;
   if (!picture || !/image/.test(picture.mimetype))
     return res.status(400).send("Picture required or Invalid picture type");
@@ -46,11 +89,33 @@ router.patch("/:id", async (req, res) => {
 
   const user = await User.findById(id);
   if (!user) return res.status(404).send("User not found");
-  fs.unlink(`./public/${user.picture}`, (err) => {
-    if (err) winston.error(err);
-  });
-  user.picture = await saveImage(picture, user.picture.split(".")[0]);
+
+  if (!/default.png/.test(user.picture))
+    fs.unlink(`./public/${user.picture}`, (err) => {
+      if (err && err.code !== "ENOENT") res.status(500).send(err);
+    });
+
+  user.picture = await saveImage(picture, `profilePic/${user._id}`);
+
   await user.save();
   res.send(user);
 });
+
+// delete user and its relations
+router.delete("/:id", async (req, res) => {
+  const id = req.params.id;
+  if (!mongoose.isValidObjectId(id))
+    return res.status(400).send("Invalid school id");
+
+  const user = await User.findByIdAndDelete(id);
+  if (!user) return res.status(404).send("User already deleted");
+  if (!/default.png/.test(user.picture))
+    fs.unlink(`./public/${user.picture}`, (err) => {
+      if (err && err.code !== "ENOENT") res.status(500).send(err);
+    });
+  const result = await deleteUser(id, user.role);
+  result.user = user;
+  res.send(result);
+});
+
 module.exports = router;
